+++ libs/components/indicators/src/lib/modules/illustration/illustration.component.ts
@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common';
 import {
   ChangeDetectionStrategy,
   Component,
@@ -18,12 +17,9 @@ import { SkyIllustrationSize } from './illustration-size';
  */
 @Component({
   selector: 'sky-illustration',
-  imports: [CommonModule],
+  imports: [],
   templateUrl: './illustration.component.html',
-  styleUrls: [
-    './illustration.default.component.scss',
-    './illustration.modern.component.scss',
-  ],
+  styleUrls: ['./illustration.component.scss'],
   hostDirectives: [SkyThemeComponentClassDirective],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
@@ -48,6 +44,15 @@ export class SkyIllustrationComponent {
     toObservable(this.name).pipe(
       switchMap((name) =>
         this.#resolverSvc ? from(this.#resolverSvc.resolveUrl(name)) : of(''),
+      ),
+      catchError(() => of('')),
+    ),
+  );
+
+  protected readonly svgHref = toSignal(
+    toObservable(this.name).pipe(
+      switchMap((name) =>
+        this.#resolverSvc ? from(this.#resolverSvc.resolveHref(name)) : of(''),
       ),
       catchError(() => of('')),
     ),